로그인 기능 구현

+++ src/main/java/backend/backend/common/refreshToken/repository/RefreshTokenRepository.java
@@ -9,5 +9,5 @@ import java.util.UUID;
 
 @Repository
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    Optional<RefreshToken> findByIdAndReissueCountLessThan(Long id, long count);
+    Optional<RefreshToken> findByIdAndReissueCountLessThan(Long id, int count);
 }
